tests_analytic/functional/requirements.txt
pytest==7.3.1
pytest-asyncio==0.21.0
pydantic==1.10.6
python-dotenv==0.21.1
backoff==2.2.1
requests==2.31.0
python-keycloak==2.15.3